sistemato c#

+++ Heat.ConvertedToC#/Controllers/InvoiceRowsController.cs
@@ -93,9 +93,9 @@ namespace Heat.Controllers
 				}
 				invoiceRowDB.Product = _db.Products.Find(invoiceRow.ProductID);
 				invoiceRowDB.Quantity = invoiceRow.Quantity;
-				invoiceRowDB.RateDiscount1 = invoiceRow.Discount1;
-				invoiceRowDB.RateDiscount2 = invoiceRow.Discount2;
-				invoiceRowDB.RateDiscount3 = invoiceRow.Discount3;
+				invoiceRowDB.RateDiscount1 =(decimal) invoiceRow.Discount1;
+				invoiceRowDB.RateDiscount2 =(decimal) invoiceRow.Discount2;
+				invoiceRowDB.RateDiscount3 = (decimal) invoiceRow.Discount3;
 				invoiceRowDB.UnitPrice = invoiceRow.UnitPrice;
 				invoiceRowDB.VAT_Rate = invoiceRow.VAT;
